:hammer: Improve the isolation of the Playground
Because the playground was a Vertical, if someone where to query (for
example) 'Vertical > Vertical' it would match the top container in the DOM,
which would be surprising given how the playground presents itself.

This commit makes a custom container and, while not making the name
impossible to guess, it does isolate it from the usual CSS types that a user
might try.