:zap: Ringfence the playground
Because of the way queries work, if someone did a query like:

    Playground > *

it would match all the widgets at the "top" level of the playground. Ideally
we want to pretend that the playground is the whole universe. So here I
prefix all queries with the playground CSS type to make it harder for there
to be a clash.